+++ Console/BudgetActions/IBudgetAction.cs
@@ -6,25 +6,3 @@ public interface IBudgetAction
    public int Order { get; }
    public void Execute(); 
 }
-
-public class ListTransactionsBudgetAction : IBudgetAction
-{
-   public string DisplayName => "List Transactions";
-   public int Order => 0;
-
-   public void Execute()
-   {
-      Console.WriteLine("Listing transactions...");
-   }
-}
-
-public class ExitBudgetAction : IBudgetAction
-{
-   public string DisplayName  => "Exit"; 
-   public int Order  => 1; 
-
-   public void Execute()
-   {
-      Environment.Exit(0);
-   }
-}
